# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RAG Tabanlı Akıllı Belge Analiz Platformu API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Document {
  id: string;
  filename: string;
  uploadedAt: string;
  pageCount: number;
  chunkCount: number;
  fileSize: number;
}

export interface DeleteResult {
  success: boolean;
  message: string;
}

export interface QueryRequest {
  question: string;
  /** Optional - if provided, query only this document */
  documentId?: string;
  topK?: number;
}

export interface SourceChunk {
  text: string;
  page: number;
  score: number;
  documentId: string;
  filename: string;
}

export interface QueryResponse {
  answer: string;
  sources: SourceChunk[];
  documentId?: string;
  question: string;
  historyId: string;
}

export interface ChatRequest {
  question: string;
  documentId?: string;
  /** Conversation session ID for maintaining context */
  sessionId?: string;
  topK?: number;
}

export interface ChatResponse {
  answer: string;
  sources: SourceChunk[];
  sessionId: string;
  question: string;
  historyId: string;
}

export interface ClearSessionRequest {
  sessionId: string;
}

export interface HistoryItem {
  id: string;
  question: string;
  answer: string;
  documentId?: string;
  documentName?: string;
  createdAt: string;
  sources: SourceChunk[];
}