# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RAG Tabanlı Akıllı Belge Analiz Platformu API
 * OpenAPI spec version: 0.1.0
 */
import type { SourceChunk } from "./sourceChunk";

export interface ChatResponse {
  answer: string;
  sources: SourceChunk[];
  sessionId: string;
  question: string;
  historyId: string;
}