# lib/api-zod/src/generated/types/clearSessionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RAG Tabanlı Akıllı Belge Analiz Platformu API
 * OpenAPI spec version: 0.1.0
 */

export interface ClearSessionRequest {
  sessionId: string;
}